dword->double - delta

+++ LBEngine/inc/Game.h
@@ -66,7 +66,7 @@ private:
 
     void CreateResources();
 
-    void Update(DWORD deltaTime);
+    void Update(double deltaTime);
     void Render();
     void Clear(const float clearColor[4], float clearDepth, UINT8 clearStencil);
     void Present();
